+++ types/signed_header.go
@@ -2,9 +2,11 @@ package types
 
 import (
 	"bytes"
+	"errors"
 	"fmt"
 
 	"github.com/celestiaorg/go-header"
+	"github.com/tendermint/tendermint/crypto/ed25519"
 	tmbytes "github.com/tendermint/tendermint/libs/bytes"
 	tmtypes "github.com/tendermint/tendermint/types"
 )
@@ -34,13 +36,13 @@ func (sH *SignedHeader) Verify(untrst header.Header) error {
 	sHHash := sH.Header.Hash()
 	if !bytes.Equal(untrstH.LastHeaderHash[:], sHHash) {
 		return &header.VerifyError{
-			Reason: fmt.Errorf("Last header hash %v does not match hash of previous header %v", untrstH.LastHeaderHash[:], sHHash),
+			Reason: fmt.Errorf("last header hash %v does not match hash of previous header %v", untrstH.LastHeaderHash[:], sHHash),
 		}
 	}
 	sHLastCommitHash := sH.getLastCommitHash()
 	if !bytes.Equal(untrstH.LastCommitHash[:], sHLastCommitHash) {
 		return &header.VerifyError{
-			Reason: fmt.Errorf("Last commit hash %v does not match hash of previous header %v", untrstH.LastCommitHash[:], sHHash),
+			Reason: fmt.Errorf("last commit hash %v does not match hash of previous header %v", untrstH.LastCommitHash[:], sHHash),
 		}
 	}
 	return sH.Header.Verify(&untrstH.Header)
@@ -65,9 +67,55 @@ func (sH *SignedHeader) getLastCommitHash() []byte {
 		lastABCICommit.Signatures = append(lastABCICommit.Signatures, commitSig)
 	}
 
+	// We don't submit a multiple signature scheme so there can only be one signature
 	if len(sH.Commit.Signatures) == 1 {
 		lastABCICommit.Signatures[0].ValidatorAddress = sH.ProposerAddress
 		lastABCICommit.Signatures[0].Timestamp = sH.Time()
 	}
 	return lastABCICommit.Hash()
+}
+
+// ValidateBasic performs basic validation of a signed header.
+func (h *SignedHeader) ValidateBasic() error {
+	err := h.Header.ValidateBasic()
+	if err != nil {
+		return err
+	}
+
+	err = h.Commit.ValidateBasic()
+	if err != nil {
+		return err
+	}
+
+	// Handle Based Rollup case
+	if h.Validators == nil || len(h.Validators.Validators) == 0 {
+		return nil
+	}
+
+	err = h.Validators.ValidateBasic()
+	if err != nil {
+		return err
+	}
+
+	if !bytes.Equal(h.Validators.Hash(), h.AggregatorsHash[:]) {
+		return errors.New("aggregator set hash in signed header and hash of validator set do not match")
+	}
+
+	// Make sure there is exactly one signature
+	if len(h.Commit.Signatures) != 1 {
+		return errors.New("expected exactly one signature")
+	}
+
+	signature := h.Commit.Signatures[0]
+	proposer := h.Validators.GetProposer()
+	var pubKey ed25519.PubKey = proposer.PubKey.Bytes()
+	msg, err := h.Header.MarshalBinary()
+	if err != nil {
+		return errors.New("signature verification failed, unable to marshal header")
+	}
+	if !pubKey.VerifySignature(msg, signature) {
+		return errors.New("signature verification failed")
+	}
+
+	return nil
 }
